requirements.txt
openai
langchain
langchain-openai
langchain-community
langchain-core
pydantic
deeplake
tiktoken
python-dotenv